moving function from main to auth
