+++ app/templates/auth/user_list.html
@@ -7,9 +7,9 @@
 
 <div class="col-xs-6 col-md-4 pull-left">
     {% if active_users %}
-        <a href="{{ url_for('main.user_list_page') }}?active=False">View Inactive Users</a>
+        <a href="{{ url_for('auth.user_list_page') }}?active=False">View Inactive Users</a>
     {% else %}
-        <a href="{{ url_for('main.user_list_page') }}?active=True">View Active Users</a>
+        <a href="{{ url_for('auth.user_list_page') }}?active=True">View Active Users</a>
     {% endif %}
 </div>
 <div clas="row">
